+++ src/neo4j/_sync/driver.py
@@ -84,7 +84,10 @@ from ..auth_management import (
     ClientCertificate,
     ClientCertificateProvider,
 )
-from ..exceptions import Neo4jError
+from ..exceptions import (
+    DriverError,
+    Neo4jError,
+)
 from .auth_management import _StaticClientCertificateProvider
 from .bookmark_manager import (
     Neo4jBookmarkManager,
@@ -555,13 +558,7 @@ class Driver:
 
     def _check_state(self):
         if self._closed:
-            # TODO: 6.0 - raise the error
-            # raise DriverError("Driver closed")
-            deprecation_warn(
-                "Using a driver after it has been closed is deprecated. "
-                "Future versions of the driver will raise an error.",
-                stack_level=3,
-            )
+            raise DriverError("Driver closed")
 
     @property
     def encrypted(self) -> bool:
@@ -613,6 +610,11 @@ class Driver:
                 key-word arguments.
 
             :returns: new :class:`neo4j.Session` object
+
+            :raises DriverError: if the driver has been closed.
+
+            .. versionchanged:: 6.0
+                Raise :exc:`DriverError` if the driver has been closed.
             """
             if "warn_notification_severity" in config:
                 # Would work just fine, but we don't want to introduce yet
@@ -650,9 +652,8 @@ class Driver:
             spawned from it (such as sessions or transactions) while calling
             this method. Failing to do so results in unspecified behavior.
         """
-        # TODO: 6.0 - NOOP if already closed
-        # if self._closed:
-        #     return
+        if self._closed:
+            return
         try:
             self._pool.close()
         except asyncio.CancelledError:
@@ -916,6 +917,8 @@ class Driver:
         :returns: the result of the ``result_transformer_``
         :rtype: T
 
+        :raises DriverError: if the driver has been closed.
+
         .. versionadded:: 5.5
 
         .. versionchanged:: 5.8
@@ -929,6 +932,9 @@ class Driver:
         .. versionchanged:: 5.15
             The ``query_`` parameter now also accepts a :class:`.Query` object
             instead of only :class:`str`.
+
+        .. versionchanged:: 6.0
+            Raise :exc:`DriverError` if the driver has been closed.
         '''  # noqa: E501 example code isn't too long
         self._check_state()
         invalid_kwargs = [
@@ -1072,11 +1078,15 @@ class Driver:
             :raises Exception: if the driver cannot connect to the remote.
                 Use the exception to further understand the cause of the
                 connectivity problem.
+            :raises DriverError: if the driver has been closed.
 
             .. versionchanged:: 5.0
                 The undocumented return value has been removed.
                 If you need information about the remote server, use
                 :meth:`get_server_info` instead.
+
+            .. versionchanged:: 6.0
+                Raise :exc:`DriverError` if the driver has been closed.
             """
             self._check_state()
             if config:
@@ -1151,8 +1161,12 @@ class Driver:
             :raises Exception: if the driver cannot connect to the remote.
                 Use the exception to further understand the cause of the
                 connectivity problem.
+            :raises DriverError: if the driver has been closed.
 
             .. versionadded:: 5.0
+
+            .. versionchanged:: 6.0
+                Raise :exc:`DriverError` if the driver has been closed.
             """
             self._check_state()
             if config:
@@ -1169,15 +1183,20 @@ class Driver:
         """
         Check if the server or cluster supports multi-databases.
 
-        :returns: Returns true if the server or cluster the driver connects to
-            supports multi-databases, otherwise false.
-
         .. note::
             Feature support query based solely on the Bolt protocol version.
             The feature might still be disabled on the server side even if this
             function return :data:`True`. It just guarantees that the driver
             won't throw a :exc:`.ConfigurationError` when trying to use this
             driver feature.
+
+        :returns: Returns true if the server or cluster the driver connects to
+            supports multi-databases, otherwise false.
+
+        :raises DriverError: if the driver has been closed.
+
+        .. versionchanged:: 6.0
+            Raise :exc:`DriverError` if the driver has been closed.
         """
         self._check_state()
         session_config = self._read_session_config({})
@@ -1245,10 +1264,14 @@ class Driver:
             :raises Exception: if the driver cannot connect to the remote.
                 Use the exception to further understand the cause of the
                 connectivity problem.
+            :raises DriverError: if the driver has been closed.
 
             .. versionadded:: 5.8
 
             .. versionchanged:: 5.14 Stabilized from experimental.
+
+            .. versionchanged:: 6.0
+                Raise :exc:`DriverError` if the driver has been closed.
             """
             self._check_state()
             if config:
@@ -1280,10 +1303,6 @@ class Driver:
         """
         Check if the remote supports connection re-authentication.
 
-        :returns: Returns true if the server or cluster the driver connects to
-            supports re-authentication of existing connections, otherwise
-            false.
-
         .. note::
             Feature support query based solely on the Bolt protocol version.
             The feature might still be disabled on the server side even if this
@@ -1291,7 +1310,16 @@ class Driver:
             won't throw a :exc:`.ConfigurationError` when trying to use this
             driver feature.
 
+        :returns: Returns true if the server or cluster the driver connects to
+            supports re-authentication of existing connections, otherwise
+            false.
+
+        :raises DriverError: if the driver has been closed.
+
         .. versionadded:: 5.8
+
+        .. versionchanged:: 6.0
+            Raise :exc:`DriverError` if the driver has been closed.
         """
         self._check_state()
         session_config = self._read_session_config({})